+++ lib/api/unittest/Main.cc
@@ -16,6 +16,7 @@
 #include "CDataFrameAnalysisRunnerTest.h"
 #include "CDataFrameAnalysisSpecificationTest.h"
 #include "CDataFrameAnalyzerTest.h"
+#include "CDataFrameAnalyzerTrainingTest.h"
 #include "CDetectionRulesJsonParserTest.h"
 #include "CFieldConfigTest.h"
 #include "CFieldDataTyperTest.h"
@@ -53,6 +54,7 @@ int main(int argc, const char** argv) {
     runner.addTest(CDataFrameAnalysisSpecificationTest::suite());
     runner.addTest(CBoostedTreeInferenceModelBuilderTest::suite());
     runner.addTest(CDataFrameAnalyzerTest::suite());
+    runner.addTest(CDataFrameAnalyzerTrainingTest::suite());
     runner.addTest(CDetectionRulesJsonParserTest::suite());
     runner.addTest(CFieldConfigTest::suite());
     runner.addTest(CFieldDataTyperTest::suite());
