+++ lib/api/unittest/Makefile
@@ -30,6 +30,7 @@ SRCS=\
 	CDataFrameAnalysisRunnerTest.cc \
 	CDataFrameAnalysisSpecificationTest.cc \
 	CDataFrameAnalyzerTest.cc \
+	CDataFrameAnalyzerTrainingTest.cc \
 	CDataFrameMockAnalysisRunner.cc \
 	CDetectionRulesJsonParserTest.cc \
 	CFieldConfigTest.cc \
